org: rafactor code
