+++ tests/test_estimate_ctf.py
@@ -2,43 +2,52 @@ import pytest
 import torch
 
 from torch_ctf_estimation.estimate_ctf import estimate_ctf
-from torch_ctf_estimation.models import LaserParams
+from torch_ctf_estimation.models import (
+    CTFFittingParams,
+    LaserParams,
+    OpticalParams,
+)
+
+
+def default_optical_params(**overrides):
+    """Build OpticalParams with test defaults; overrides merged in."""
+    p = {
+        "pixel_spacing_angstroms": 1.0,
+        "voltage_kev": 300.0,
+        "spherical_aberration_mm": 2.7,
+        "amplitude_contrast_fraction": 0.1,
+    }
+    p.update(overrides)
+    return OpticalParams(**p)
+
+
+def default_fitting_params(**overrides):
+    """Build CTFFittingParams with test defaults; overrides merged in."""
+    p = {
+        "defocus_grid_resolution": (1, 2, 2),
+        "frequency_fit_range_angstroms": (30.0, 5.0),
+        "defocus_range_microns": (0.5, 3.0),
+        "patch_sidelength": 128,
+    }
+    p.update(overrides)
+    return CTFFittingParams(**p)
 
 
 def test_estimate_ctf_2d_image():
     """Test estimate_ctf with a 2D image."""
-    # Create a synthetic 2D image
     image = torch.randn(1024, 1024)
-
-    # Define typical cryo-EM parameters
-    pixel_spacing_angstroms = 1.0
-    defocus_grid_resolution = (1, 3, 3)  # (t, h, w)
-    frequency_fit_range_angstroms = (30.0, 5.0)  # (low, high)
-    defocus_range_microns = (0.5, 5.0)  # (low, high)
-    voltage_kev = 300.0
-    spherical_aberration_mm = 2.7
-    amplitude_contrast_fraction = 0.1
-    patch_sidelength = 128
-
-    # Run estimation (returns mean_ps, result1d, result2d)
-    _mean_ps, _result1d, result2d = estimate_ctf(
-        image=image,
-        pixel_spacing_angstroms=pixel_spacing_angstroms,
-        defocus_grid_resolution=defocus_grid_resolution,
-        frequency_fit_range_angstroms=frequency_fit_range_angstroms,
-        defocus_range_microns=defocus_range_microns,
-        voltage_kev=voltage_kev,
-        spherical_aberration_mm=spherical_aberration_mm,
-        amplitude_contrast_fraction=amplitude_contrast_fraction,
-        patch_sidelength=patch_sidelength,
+    optical = default_optical_params()
+    fitting = default_fitting_params(
+        defocus_grid_resolution=(1, 3, 3),
+        defocus_range_microns=(0.5, 5.0),
+        patch_sidelength=128,
     )
-
-    # Default is grid model
+    _mean_ps, _result1d, result2d = estimate_ctf(
+        image, optical, fitting, device=torch.device("cpu")
+    )
     assert result2d.defocus_model_type == "grid"
-    # (defocus_model.data has shape (1, t, h, w), squeeze to (t, h, w))
     defocus_field = result2d.defocus_model.data.squeeze(0)
-    expected_shape = defocus_grid_resolution
-    assert defocus_field.shape == expected_shape
+    assert defocus_field.shape == (1, 3, 3)
 
     # Check defocus values are within reasonable range
     # assert torch.all(defocus_field >= defocus_range_microns[0])
@@ -48,28 +57,15 @@ def test_estimate_ctf_2d_image():
 def test_estimate_ctf_2d_image_linear_model():
     """Test estimate_ctf with 2D image and linear defocus model."""
     image = torch.randn(1024, 1024)
-    pixel_spacing_angstroms = 1.0
-    defocus_grid_resolution = (1, 1, 1)  # only nt=1 used for linear
-    frequency_fit_range_angstroms = (30.0, 5.0)
-    defocus_range_microns = (0.5, 5.0)
-    voltage_kev = 300.0
-    spherical_aberration_mm = 2.7
-    amplitude_contrast_fraction = 0.1
-    patch_sidelength = 128
-
-    _mean_ps, _result1d, result2d = estimate_ctf(
-        image=image,
-        pixel_spacing_angstroms=pixel_spacing_angstroms,
-        defocus_grid_resolution=defocus_grid_resolution,
-        frequency_fit_range_angstroms=frequency_fit_range_angstroms,
-        defocus_range_microns=defocus_range_microns,
-        voltage_kev=voltage_kev,
-        spherical_aberration_mm=spherical_aberration_mm,
-        amplitude_contrast_fraction=amplitude_contrast_fraction,
-        patch_sidelength=patch_sidelength,
+    optical = default_optical_params()
+    fitting = default_fitting_params(
+        defocus_grid_resolution=(1, 1, 1),
+        defocus_range_microns=(0.5, 5.0),
         defocus_model="linear",
     )
-
+    _mean_ps, _result1d, result2d = estimate_ctf(
+        image, optical, fitting, device=torch.device("cpu")
+    )
     assert result2d.defocus_model_type == "linear"
     linear = result2d.defocus_model
     assert hasattr(linear, "defocus_0")
@@ -81,37 +77,25 @@ def test_estimate_ctf_2d_image_linear_model():
 
 def test_estimate_ctf_3d_image():
     """Test estimate_ctf with a 3D image stack."""
-    # Create a synthetic 3D image stack
     image = torch.randn(4, 256, 256)
-
-    # Define typical cryo-EM parameters
-    pixel_spacing_angstroms = 1.5
-    defocus_grid_resolution = (4, 2, 2)  # (t, h, w)
-    frequency_fit_range_angstroms = (20.0, 4.0)  # (low, high)
-    defocus_range_microns = (1.0, 4.0)  # (low, high)
-    voltage_kev = 200.0
-    spherical_aberration_mm = 2.0
-    amplitude_contrast_fraction = 0.07
-    patch_sidelength = 64
-
-    # Run estimation (returns mean_ps, result1d, result2d)
-    _mean_ps, _result1d, result2d = estimate_ctf(
-        image=image,
-        pixel_spacing_angstroms=pixel_spacing_angstroms,
-        defocus_grid_resolution=defocus_grid_resolution,
-        frequency_fit_range_angstroms=frequency_fit_range_angstroms,
-        defocus_range_microns=defocus_range_microns,
-        voltage_kev=voltage_kev,
-        spherical_aberration_mm=spherical_aberration_mm,
-        amplitude_contrast_fraction=amplitude_contrast_fraction,
-        patch_sidelength=patch_sidelength,
+    optical = default_optical_params(
+        pixel_spacing_angstroms=1.5,
+        voltage_kev=200.0,
+        spherical_aberration_mm=2.0,
+        amplitude_contrast_fraction=0.07,
     )
-
+    fitting = default_fitting_params(
+        defocus_grid_resolution=(4, 2, 2),
+        frequency_fit_range_angstroms=(20.0, 4.0),
+        defocus_range_microns=(1.0, 4.0),
+        patch_sidelength=64,
+    )
+    _mean_ps, _result1d, result2d = estimate_ctf(
+        image, optical, fitting, device=torch.device("cpu")
+    )
     assert result2d.defocus_model_type == "grid"
-    # (defocus_model.data has shape (1, t, h, w), squeeze to (t, h, w))
     defocus_field = result2d.defocus_model.data.squeeze(0)
-    expected_shape = defocus_grid_resolution
-    assert defocus_field.shape == expected_shape
+    assert defocus_field.shape == (4, 2, 2)
 
     # Check defocus values are within reasonable range
     # assert torch.all(defocus_field >= defocus_range_microns[0])
@@ -121,100 +105,74 @@ def test_estimate_ctf_3d_image():
 def test_estimate_ctf_whole_image_mode():
     """Test estimate_ctf with whole-image mode (patch_sidelength < 0)."""
     image = torch.randn(256, 256)
-    pixel_spacing_angstroms = 1.0
-    defocus_grid_resolution = (1, 1, 1)
-    frequency_fit_range_angstroms = (30.0, 5.0)
-    defocus_range_microns = (0.5, 5.0)
-    voltage_kev = 300.0
-    spherical_aberration_mm = 2.7
-    amplitude_contrast_fraction = 0.1
-
-    mean_ps, _result1d, result2d = estimate_ctf(
-        image=image,
-        pixel_spacing_angstroms=pixel_spacing_angstroms,
-        defocus_grid_resolution=defocus_grid_resolution,
-        frequency_fit_range_angstroms=frequency_fit_range_angstroms,
-        defocus_range_microns=defocus_range_microns,
-        voltage_kev=voltage_kev,
-        spherical_aberration_mm=spherical_aberration_mm,
-        amplitude_contrast_fraction=amplitude_contrast_fraction,
+    optical = default_optical_params()
+    fitting = default_fitting_params(
+        defocus_grid_resolution=(1, 1, 1),
+        defocus_range_microns=(0.5, 5.0),
         patch_sidelength=-1,
     )
-
+    mean_ps, _result1d, result2d = estimate_ctf(
+        image, optical, fitting, device=torch.device("cpu")
+    )
     assert mean_ps.dim() == 2
     assert result2d.defocus_model_type == "grid"
     defocus_field = result2d.defocus_model.data.squeeze(0)
-    assert defocus_field.shape == defocus_grid_resolution
+    assert defocus_field.shape == (1, 1, 1)
 
 
 def test_estimate_ctf_whole_image_mode_rejects_nh_nw_not_1():
     """Test that whole-image mode raises when nh or nw is not 1."""
     image = torch.randn(256, 256)
+    optical = default_optical_params()
     with pytest.raises(ValueError, match="nh=1 and nw=1"):
         estimate_ctf(
-            image=image,
-            pixel_spacing_angstroms=1.0,
-            defocus_grid_resolution=(1, 2, 1),
-            frequency_fit_range_angstroms=(30.0, 5.0),
-            defocus_range_microns=(0.5, 5.0),
-            voltage_kev=300.0,
-            spherical_aberration_mm=2.7,
-            amplitude_contrast_fraction=0.1,
-            patch_sidelength=-1,
+            image,
+            optical,
+            default_fitting_params(
+                defocus_grid_resolution=(1, 2, 1),
+                patch_sidelength=-1,
+            ),
         )
     with pytest.raises(ValueError, match="nh=1 and nw=1"):
         estimate_ctf(
-            image=image,
-            pixel_spacing_angstroms=1.0,
-            defocus_grid_resolution=(1, 1, 2),
-            frequency_fit_range_angstroms=(30.0, 5.0),
-            defocus_range_microns=(0.5, 5.0),
-            voltage_kev=300.0,
-            spherical_aberration_mm=2.7,
-            amplitude_contrast_fraction=0.1,
-            patch_sidelength=-1,
+            image,
+            optical,
+            default_fitting_params(
+                defocus_grid_resolution=(1, 1, 2),
+                patch_sidelength=-1,
+            ),
         )
 
 
 def test_estimate_ctf_use_1d_defocus_for_spatial():
     """Test use_1d_defocus_for_spatial returns grid/linear result with correct shape."""
     image = torch.randn(512, 512)
-    defocus_grid_resolution = (1, 2, 2)
-    _mean_ps, _result1d, result2d = estimate_ctf(
-        image=image,
-        pixel_spacing_angstroms=1.0,
-        defocus_grid_resolution=defocus_grid_resolution,
-        frequency_fit_range_angstroms=(30.0, 5.0),
-        defocus_range_microns=(0.5, 5.0),
-        voltage_kev=300.0,
-        spherical_aberration_mm=2.7,
-        amplitude_contrast_fraction=0.1,
-        patch_sidelength=128,
+    optical = default_optical_params()
+    fitting = default_fitting_params(
+        defocus_grid_resolution=(1, 2, 2),
         use_1d_defocus_for_spatial=True,
         defocus_model="grid",
     )
+    _mean_ps, _result1d, result2d = estimate_ctf(
+        image, optical, fitting, device=torch.device("cpu")
+    )
     assert result2d.defocus_model_type == "grid"
     defocus_field = result2d.defocus_model.data.squeeze(0)
-    assert defocus_field.shape == defocus_grid_resolution
+    assert defocus_field.shape == (1, 2, 2)
     assert result2d.astigmatism is not None or result2d.astigmatism is None
 
 
 def test_estimate_ctf_use_1d_defocus_for_spatial_linear():
     """Test use_1d_defocus_for_spatial with linear model."""
     image = torch.randn(512, 512)
-    defocus_grid_resolution = (1, 2, 2)
-    _mean_ps, _result1d, result2d = estimate_ctf(
-        image=image,
-        pixel_spacing_angstroms=1.0,
-        defocus_grid_resolution=defocus_grid_resolution,
-        frequency_fit_range_angstroms=(30.0, 5.0),
-        defocus_range_microns=(0.5, 5.0),
-        voltage_kev=300.0,
-        spherical_aberration_mm=2.7,
-        amplitude_contrast_fraction=0.1,
-        patch_sidelength=128,
+    optical = default_optical_params()
+    fitting = default_fitting_params(
+        defocus_grid_resolution=(1, 2, 2),
         use_1d_defocus_for_spatial=True,
         defocus_model="linear",
+    )
+    _mean_ps, _result1d, result2d = estimate_ctf(
+        image, optical, fitting, device=torch.device("cpu")
     )
     assert result2d.defocus_model_type == "linear"
     assert hasattr(result2d.defocus_model, "defocus_0")
@@ -225,19 +183,14 @@ def test_estimate_ctf_use_1d_defocus_for_spatial_linear():
 def test_estimate_ctf_linear_fix_defocus_0_from_1x1():
     """Test linear_fix_defocus_0_from_1x1: defocus_0 comes from 2D@1x1."""
     image = torch.randn(512, 512)
-    defocus_grid_resolution = (1, 2, 2)
-    _mean_ps, _result1d, result2d = estimate_ctf(
-        image=image,
-        pixel_spacing_angstroms=1.0,
-        defocus_grid_resolution=defocus_grid_resolution,
-        frequency_fit_range_angstroms=(30.0, 5.0),
-        defocus_range_microns=(0.5, 5.0),
-        voltage_kev=300.0,
-        spherical_aberration_mm=2.7,
-        amplitude_contrast_fraction=0.1,
-        patch_sidelength=128,
+    optical = default_optical_params()
+    fitting = default_fitting_params(
+        defocus_grid_resolution=(1, 2, 2),
         defocus_model="linear",
         linear_fix_defocus_0_from_1x1=True,
+    )
+    _mean_ps, _result1d, result2d = estimate_ctf(
+        image, optical, fitting, device=torch.device("cpu")
     )
     assert result2d.defocus_model_type == "linear"
     assert result2d.defocus_model.defocus_0 is not None
@@ -248,19 +201,14 @@ def test_estimate_ctf_linear_fix_defocus_0_from_1x1():
 def test_estimate_ctf_linear_fix_defocus_0_2d_zncc():
     """Test linear_fix_defocus_0_from_1x1 with gradient from 2D ZNCC."""
     image = torch.randn(512, 512)
-    defocus_grid_resolution = (1, 2, 2)
-    _mean_ps, _result1d, result2d = estimate_ctf(
-        image=image,
-        pixel_spacing_angstroms=1.0,
-        defocus_grid_resolution=defocus_grid_resolution,
-        frequency_fit_range_angstroms=(30.0, 5.0),
-        defocus_range_microns=(0.5, 5.0),
-        voltage_kev=300.0,
-        spherical_aberration_mm=2.7,
-        amplitude_contrast_fraction=0.1,
-        patch_sidelength=128,
+    optical = default_optical_params()
+    fitting = default_fitting_params(
+        defocus_grid_resolution=(1, 2, 2),
         defocus_model="linear",
         linear_fix_defocus_0_from_1x1=True,
+    )
+    _mean_ps, _result1d, result2d = estimate_ctf(
+        image, optical, fitting, device=torch.device("cpu")
     )
     assert result2d.defocus_model_type == "linear"
     assert result2d.defocus_model.defocus_0 is not None
@@ -272,16 +220,10 @@ def test_estimate_ctf_phase_shift_default_zero():
     """Test that without optimize_phase_shift, phase_shift_degrees is 0."""
     image = torch.randn(512, 512)
     _mean_ps, result1d, result2d = estimate_ctf(
-        image=image,
-        pixel_spacing_angstroms=1.0,
-        defocus_grid_resolution=(1, 2, 2),
-        frequency_fit_range_angstroms=(30.0, 5.0),
-        defocus_range_microns=(0.5, 3.0),
-        voltage_kev=300.0,
-        spherical_aberration_mm=2.7,
-        amplitude_contrast_fraction=0.1,
-        patch_sidelength=128,
-        device="cpu",
+        image,
+        default_optical_params(),
+        default_fitting_params(),
+        device=torch.device("cpu"),
     )
     phase_1d = result1d.ctf_model.phase_shift_degrees
     assert phase_1d is not None
@@ -295,18 +237,15 @@ def test_estimate_ctf_phase_shift_default_zero():
 def test_estimate_ctf_optimize_phase_shift_1d():
     """Test 1D with optimize_phase_shift=True; phase in [0, 90] (folded from 0-180)."""
     image = torch.randn(512, 512)
-    _mean_ps, result1d, _result2d = estimate_ctf(
-        image=image,
-        pixel_spacing_angstroms=1.0,
+    fitting = default_fitting_params(
         defocus_grid_resolution=(1, 1, 1),
-        frequency_fit_range_angstroms=(30.0, 5.0),
-        defocus_range_microns=(0.5, 3.0),
-        voltage_kev=300.0,
-        spherical_aberration_mm=2.7,
-        amplitude_contrast_fraction=0.1,
-        patch_sidelength=128,
         optimize_phase_shift=True,
-        device="cpu",
+    )
+    _mean_ps, result1d, _result2d = estimate_ctf(
+        image,
+        default_optical_params(),
+        fitting,
+        device=torch.device("cpu"),
     )
     phase = result1d.ctf_model.phase_shift_degrees
     assert phase is not None
@@ -317,19 +256,15 @@ def test_estimate_ctf_optimize_phase_shift_1d():
 def test_estimate_ctf_optimize_phase_shift_2d_grid():
     """Test 2D with optimize_phase_shift=True and phase_shift_model='grid'."""
     image = torch.randn(512, 512)
-    _mean_ps, _, result2d = estimate_ctf(
-        image=image,
-        pixel_spacing_angstroms=1.0,
-        defocus_grid_resolution=(1, 2, 2),
-        frequency_fit_range_angstroms=(30.0, 5.0),
-        defocus_range_microns=(0.5, 3.0),
-        voltage_kev=300.0,
-        spherical_aberration_mm=2.7,
-        amplitude_contrast_fraction=0.1,
-        patch_sidelength=128,
+    fitting = default_fitting_params(
         optimize_phase_shift=True,
         phase_shift_model="grid",
-        device="cpu",
+    )
+    _mean_ps, _, result2d = estimate_ctf(
+        image,
+        default_optical_params(),
+        fitting,
+        device=torch.device("cpu"),
     )
     assert result2d.phase_shift_degrees is not None
     assert 0.0 <= result2d.phase_shift_degrees <= 90.0
@@ -343,19 +278,15 @@ def test_estimate_ctf_optimize_phase_shift_2d_grid():
 def test_estimate_ctf_optimize_phase_shift_2d_quadratic():
     """Test 2D with optimize_phase_shift=True and phase_shift_model='quadratic'."""
     image = torch.randn(512, 512)
-    _mean_ps, _, result2d = estimate_ctf(
-        image=image,
-        pixel_spacing_angstroms=1.0,
-        defocus_grid_resolution=(1, 2, 2),
-        frequency_fit_range_angstroms=(30.0, 5.0),
-        defocus_range_microns=(0.5, 3.0),
-        voltage_kev=300.0,
-        spherical_aberration_mm=2.7,
-        amplitude_contrast_fraction=0.1,
-        patch_sidelength=128,
+    fitting = default_fitting_params(
         optimize_phase_shift=True,
         phase_shift_model="quadratic",
-        device="cpu",
+    )
+    _mean_ps, _, result2d = estimate_ctf(
+        image,
+        default_optical_params(),
+        fitting,
+        device=torch.device("cpu"),
     )
     assert result2d.phase_shift_degrees is not None
     # Allow small negative (wrap-around when quadratic C is near 180°)
@@ -376,16 +307,10 @@ def test_estimate_ctf_raises_when_rescaled_image_smaller_than_patch():
         ValueError, match=r"Rescaled image size.*smaller than patch_sidelength"
     ):
         estimate_ctf(
-            image=image,
-            pixel_spacing_angstroms=1.0,
-            defocus_grid_resolution=(1, 2, 2),
-            frequency_fit_range_angstroms=(30.0, 5.0),
-            defocus_range_microns=(0.5, 3.0),
-            voltage_kev=300.0,
-            spherical_aberration_mm=2.7,
-            amplitude_contrast_fraction=0.1,
-            patch_sidelength=128,
-            device="cpu",
+            image,
+            default_optical_params(),
+            default_fitting_params(patch_sidelength=128),
+            device=torch.device("cpu"),
         )
 
 
@@ -393,17 +318,11 @@ def test_estimate_ctf_2d_default_no_laser():
     """estimate_ctf with laser_params=None uses normal CTF and returns valid result."""
     image = torch.randn(512, 512)
     _mean_ps, _result1d, result2d = estimate_ctf(
-        image=image,
-        pixel_spacing_angstroms=1.0,
-        defocus_grid_resolution=(1, 2, 2),
-        frequency_fit_range_angstroms=(30.0, 5.0),
-        defocus_range_microns=(0.5, 5.0),
-        voltage_kev=300.0,
-        spherical_aberration_mm=2.7,
-        amplitude_contrast_fraction=0.1,
-        patch_sidelength=128,
+        image,
+        default_optical_params(),
+        default_fitting_params(defocus_range_microns=(0.5, 5.0)),
         laser_params=None,
-        device="cpu",
+        device=torch.device("cpu"),
     )
     assert result2d.defocus_model_type == "grid"
     assert result2d.defocus_model.data.shape[1:] == (1, 2, 2)
@@ -412,19 +331,12 @@ def test_estimate_ctf_2d_default_no_laser():
 def test_estimate_ctf_2d_with_laser_params():
     """estimate_ctf with laser_params set uses LPP CTF and returns Defocus2DResults."""
     image = torch.randn(512, 512)
-    laser_params = LaserParams()
     _mean_ps, _result1d, result2d = estimate_ctf(
-        image=image,
-        pixel_spacing_angstroms=1.0,
-        defocus_grid_resolution=(1, 2, 2),
-        frequency_fit_range_angstroms=(30.0, 5.0),
-        defocus_range_microns=(0.5, 5.0),
-        voltage_kev=300.0,
-        spherical_aberration_mm=2.7,
-        amplitude_contrast_fraction=0.1,
-        patch_sidelength=128,
-        laser_params=laser_params,
-        device="cpu",
+        image,
+        default_optical_params(),
+        default_fitting_params(defocus_range_microns=(0.5, 5.0)),
+        laser_params=LaserParams(),
+        device=torch.device("cpu"),
     )
     assert result2d.defocus_model_type == "grid"
     assert result2d.defocus_model.data.shape[1:] == (1, 2, 2)
